--- src/Applications.tsx
@@ -1,5 +1,0 @@
-function Applications() {
-  return <div></div>;
-}
-
-export default Applications;
