--- src/Approved.tsx
@@ -1,5 +1,0 @@
-function Approved() {
-  return <div></div>;
-}
-
-export default Approved;
