+++ src/Dashboard.tsx
@@ -2,7 +2,8 @@ import { useEffect, useState } from "react";
 import { getAppplicants } from "./services/fakeApplicantsService";
 import IApplicant from "./types/IApplicant";
 import Table from "./components/Table";
-import applicantsContext from "./context/ApplicantsContext";
+import applicantsContext from "./Context/ApplicantsContext";
+import Sidebar from "./Sidebar";
 
 function Items() {
   const [applicants, setApplicants] = useState<IApplicant[]>([]);
@@ -14,6 +15,7 @@ function Items() {
   return (
     <>
       <applicantsContext.Provider value={applicants}>
+        <Sidebar />
         <Table />
       </applicantsContext.Provider>
     </>
