--- src/NotApproved.tsx
@@ -1,5 +1,0 @@
-function NotApproved() {
-  return <div></div>;
-}
-
-export default NotApproved;
