--- src/SavedProfiles.tsx
@@ -1,5 +1,0 @@
-function SavedProfiles() {
-  return <div></div>;
-}
-
-export default SavedProfiles;
