--- src/Selections.tsx
@@ -1,5 +1,0 @@
-function Selections() {
-  return <div></div>;
-}
-
-export default Selections;
