--- src/Step1.tsx
@@ -1,5 +1,0 @@
-function Step1() {
-  return <div></div>;
-}
-
-export default Step1;
