--- src/Step2.tsx
@@ -1,5 +1,0 @@
-function Step2() {
-  return <div></div>;
-}
-
-export default Step2;
