+++ src/components/Table.tsx
@@ -1,13 +1,8 @@
-import TableBody from "./common/TableBody";
-import TableHeader from "./common/TableHeader";
+import TableBody from "../common/TableBody";
+import TableHeader from "../common/TableHeader";
 import styled from "styled-components";
-import { useContext } from "react";
-import IApplicant from "../types/IApplicant";
-import applicantsContext from "../context/ApplicantsContext";
 
 function Table() {
-  const applicants = useContext(applicantsContext) as IApplicant[];
-
   return (
     <>
       <TableCSS>
